% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAxisLimits.R
\name{calcAxisLimits}
\alias{calcAxisLimits}
\title{calculate the Y-limits for use when plotting the data}
\usage{
calcAxisLimits(inputLocation, assay = "")
}
\arguments{
\item{inputLocation, }{the full pathname of the data directory}

\item{assay, }{the name you wish assigned to the set of data you are analysis,}
}
\value{
create file limisFile.txt that will be used to set Y limits in figures
}
\description{
calculate the Y-limits for use when plotting the data
}
