% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtyCheck.R
\name{dirtyCheck}
\alias{dirtyCheck}
\title{determine if the data or the fitting process is outside of expected ranges}
\usage{
dirtyCheck(L, Y, EC50, goodness, antibody, measure, vehicle, dsname)
}
\arguments{
\item{L, }{data (x-axis)}

\item{Y, }{data (y-axis)}

\item{EC50, }{}

\item{goodness, }{}

\item{antibody, }{}

\item{measure, }{}

\item{vehicle, }{}

\item{dsname}{}
}
\value{

}
\description{
determine if the data or the fitting process is outside of expected ranges
}
