% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenameSplit.R
\name{filenameSplit}
\alias{filenameSplit}
\title{splits filename into its component parts}
\usage{
filenameSplit(filename)
}
\arguments{
\item{filename, }{}
}
\value{
date,donor,plate,control and sample no.
}
\description{
splits filename into its component parts
}
