% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEC50.R
\name{fitEC50}
\alias{fitEC50}
\title{Fit a Hill curve to the data}
\usage{
fitEC50(L, Y, measure)
}
\arguments{
\item{L, }{data (x-axis)}

\item{Y, }{data (y-axis)}

\item{measure, }{ppositive or MedianFI}
}
\value{
The ec50, the hill coefficient and goodness of fit
}
\description{
Fit a Hill curve to the data
}
