% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBarcode.R
\name{processBarcode}
\alias{processBarcode}
\title{Function to produce a barcode for each donor (and a summary csv file),}
\usage{
processBarcode(inputFile, outputLocation, dsname = "", singlePointInfo = "No")
}
\arguments{
\item{outputLocation, }{pathname where generated figures are created}

\item{dsname, }{a cohort name that will be added to all files}

\item{singlePointInfo, }{Yes or No (default), generate data describing response to a single concentration
of an agonist (this requires the data to have been processed in step 1 and 2)}

\item{inputName, }{the full pathname of the file that stores summary data}
}
\value{
create figures in outputLocation
}
\description{
Function to produce a barcode for each donor (and a summary csv file),
}
