% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{process raw data, Step D1: creates plots for each donor and a summary file}
\usage{
processData(inputLoc, outputLoc, dsname = "", singlePointInfo = "none")
}
\arguments{
\item{inputLoc, }{the full pathname of the data directory}

\item{outputLoc, }{pathname to store the results (figures and summary file)}

\item{dsname, }{a cohort name added to all files}

\item{singlePointInfo, }{extract the concentration acheived at a single point and add it to the summary file (this requires a file 'singlePointFile.txt' in the data directory)}
}
\value{
create figures and summary files in outputLocation
}
\description{
process raw data, Step D1: creates plots for each donor and a summary file
}
