% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSummary.R
\name{processSummary}
\alias{processSummary}
\title{Process summary data (step D2),}
\usage{
processSummary(inputFile, outputLocation, dsname = "", singlePointInfo = "No")
}
\arguments{
\item{inputFile, }{the pathname of the summary file to be processed}

\item{outputLocation, }{pathname of directory where generated figures are stored}

\item{dsname, }{a cohort name added to all generated files}

\item{singlePointInfo, }{Yes or No (default), generate data describing response to a single concentration
of an agonist (this requires the data to have been processed in step 1)}
}
\value{
creates figures and a summary file in the outputLocation

# processSummary('../outputs/CurveFitting/summaryFileMethodsV1.csv','../outputs/SummaryStats','Methods','Yes')
}
\description{
Process summary data (step D2),
}
